# optimize cat file batch info writing