When cat-file --batch or --batch-check is used, we can skip having to expand the format if no format is specified or if the default format is specified. In this case we know exactly how to print the objects without the full expanded format.

This was first discussed in [1].

We get a little performance boost from this optimization because this happens for each objects provided to --batch, --batch-check, or --batch-command. Because batch_object_write() is called on every oid provided in batch mode, this optimization adds up when a large number of oid info is printed.

git rev-list --all >/tmp/all-objs.txt

git cat-file --batch-check </tmp/all-obj.txt (with hyperfine)

run on origin/master:

Time (mean ± σ):      57.6 ms ±   1.7 ms    [User: 51.5 ms, System: 6.2 ms]
Range (min … max):    54.6 ms …  64.7 ms    50 runs

run on jc/optimize-cat-file-batch-default-format:

Time (mean ± σ):      49.8 ms ±   1.7 ms    [User: 42.6 ms, System: 7.3 ms]
Range (min … max):    46.9 ms …  55.9 ms    56 runs

Changes since v3:
- reuse scratch string buffer
- use --batch-all-objects in test

Changes since v2:
- get rid of type_name string buffer, and call type_name() for each object in print_default_format()

Changes since v1:
- set opt->format in batch_objects so that the loop that prints objects only has to check if the format is null to know to print the object info in the default format
- fixed up commit trailer to include Ævar as Signed-off-by

1. https://lore.kernel.org/git/87eecf8ork.fsf@evledraar.gmail.com/

cc: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
cc: Taylor Blau <me@ttaylorr.com>